requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Assets==2.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
webassets==2.0
Werkzeug==3.0.3
WTForms==3.1.2
